B33rn4ryReader: fix reading / decoding of tagID for serial-reader
as per datashee we receive the TagID and a checksum from the reader,
like "data:F004414A2E; cksum:D". The first and last byte of data seem
rubbish, so we keep "data:004414A2". This hex-value we just need to
convert to integer to get our ID (as read from the USB-reader).

Using the correct values also seems to make the checksum-code working
as the fail messages are gone.